# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST API for eSport kit App",
    "title": "eSport kit",
    "contact": {},
    "version": "1.1"
  },
  "host": "localhost:5000",
  "basePath": "/api/v1/",
  "paths": {
    "/auth/login": {
      "post": {
        "description": "user sign in",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "summary": "Learner Login",
        "parameters": [
          {
            "description": "login info",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.UserLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.TokenResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "description": "create user account",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "summary": "User Register",
        "parameters": [
          {
            "description": "register info",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.UserRegister"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/v1.idResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          }
        }
      }
    },
    "/gym/create": {
      "post": {
        "security": [
          {
            "AdminAuth": []
          }
        ],
        "description": "creating a gym",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "gym"
        ],
        "summary": "Create Gym",
        "parameters": [
          {
            "description": "gym info",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.CreateGym"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/v1.idResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          }
        }
      }
    },
    "/user/refresh-tokens": {
      "post": {
        "description": "user refresh tokens",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "user"
        ],
        "summary": "User Refresh Tokens",
        "parameters": [
          {
            "description": "register info",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.RefreshToken"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.TokenResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          }
        }
      }
    },
    "/user/verify/{code}": {
      "post": {
        "security": [
          {
            "UserAuth": []
          }
        ],
        "description": "user verify registration",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "user"
        ],
        "summary": "User Verify Registration",
        "parameters": [
          {
            "type": "string",
            "description": "verification code",
            "name": "code",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.TokenResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          },
          "default": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/v1.response"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "dto.CreateGym": {
      "type": "object",
      "required": [
        "address",
        "title"
      ],
      "properties": {
        "address": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "dto.RefreshToken": {
      "type": "object",
      "required": [
        "token"
      ],
      "properties": {
        "token": {
          "type": "string"
        }
      }
    },
    "dto.TokenResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        }
      }
    },
    "dto.UserLogin": {
      "type": "object",
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "maxLength": 64
        },
        "password": {
          "type": "string",
          "maxLength": 64,
          "minLength": 1
        }
      }
    },
    "dto.UserRegister": {
      "type": "object",
      "required": [
        "email",
        "name",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "maxLength": 63
        },
        "name": {
          "type": "string",
          "maxLength": 64,
          "minLength": 1
        },
        "password": {
          "type": "string",
          "maxLength": 64,
          "minLength": 1
        }
      }
    },
    "v1.idResponse": {
      "type": "object",
      "properties": {
        "id": {}
      }
    },
    "v1.response": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "AdminAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    },
    "TrainerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    },
    "UserAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}